feat(layout.tsx, sticky-header.tsx): add aria-label to improve accessibility for home link refactor(page.tsx): move getPageJsonLd and Pattern functions to the bottom for better organization fix(award-item.tsx, certification-item.tsx, project-item.tsx): add aria-hidden to decorative divs for accessibility refactor(email-item.tsx, overview/index.tsx, job-item.tsx, phone-item.tsx): change icon prop to use component type instead of JSX element refactor(intro-item.tsx): update icon prop to use component type and add aria-hidden to icon container feat(teck-stack.tsx): consolidate tech stack components into a single file for simplicity and maintainability
