+++ src/features/profile/components/awards/award-item.tsx
@@ -21,7 +21,10 @@ export function AwardItem({
   return (
     <AccordionPrimitive.Item value={award.id} disabled={!canExpand} asChild>
       <div className={cn("flex items-center", className)}>
-        <div className="mx-4 flex size-6 shrink-0 items-center justify-center rounded-lg border bg-zinc-50 text-muted-foreground shadow-xs dark:bg-zinc-900">
+        <div
+          className="mx-4 flex size-6 shrink-0 items-center justify-center rounded-lg border bg-zinc-50 text-muted-foreground shadow-xs dark:bg-zinc-900"
+          aria-hidden="true"
+        >
           <AwardIcon className="size-4" />
         </div>
