+++ src/features/profile/components/certifications/certification-item.tsx
@@ -26,7 +26,10 @@ export function CertificationItem({
           className="mx-4 flex size-6 shrink-0"
         />
       ) : (
-        <div className="mx-4 flex size-6 shrink-0 items-center justify-center rounded-lg border bg-zinc-50 text-muted-foreground shadow-xs dark:bg-zinc-900">
+        <div
+          className="mx-4 flex size-6 shrink-0 items-center justify-center rounded-lg border bg-zinc-50 text-muted-foreground shadow-xs dark:bg-zinc-900"
+          aria-hidden="true"
+        >
           <CheckCircleIcon className="size-4" />
         </div>
       )}
